+++ PKHeX.Core/Legality/Evolutions/EvolutionSet.cs
@@ -1,0 +1,10 @@
+namespace PKHeX.Core
+{
+    /// <summary>
+    /// Table of Evolution Branch Entries
+    /// </summary>
+    public abstract class EvolutionSet
+    {
+        public EvolutionMethod[] PossibleEvolutions;
+    }
+}
