+++ PKHeX.Core/Legality/Evolutions/EvolutionStage.cs
@@ -1,0 +1,12 @@
+using System.Collections.Generic;
+
+namespace PKHeX.Core
+{
+    /// <summary>
+    /// Evolution Stage Entries
+    /// </summary>
+    public struct EvolutionStage
+    {
+        public List<EvolutionMethod> StageEntryMethods;
+    }
+}
